% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots map of the state with accident data for a specified year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a numeric value}

\item{year}{a character string or numeric}
}
\value{
plots a map for data filtered for the inputed state number from the file where filename contains
the inputted year. If the state number is not found the the function will report an error message "invalid STATE
number". If not accident records found for the state then a message is returned "no accident to plot" with a return
value NULL
}
\description{
This function returns a FARS report for the state number and year passed as an argument
}
\note{
The function replace the Longitude data with NA where the Logitude value is > 900
the function replace the Latitude data with NA where the Latitute value is > 90
}
