% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Extract Months and years of the FARS files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a character string or numeric or a list of character stings/ numeric specifying year/ years}
}
\value{
a list of months and years from the FARS files for the inputted \code{years}.
}
\description{
This function take year or list of years as aurgument and returns list of yearwise months and years from the FARS
file for the inputted years.
}
\note{
The function throws a warning of invalid year if the file with year in the name is not found. Also
 must be in the current working directory


#@examples
#fars_read_years("2013")
#fars_read_years(2013)
#fars_read_years(c(2012, 2013, 2014, 2015))
#fars_read_years(c("2012", "2013", "2014", "2015"))
}
