% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Monthwise and yearwise summary of FARS data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a character string or numeric or a list of character stings/ numeric specifying year/ years}
}
\value{
a table summarizing the FARS data monthwise for the each inputted year.
}
\description{
This function provide the summary of count for each month of an year for the years passed as
argument.
}
\note{
The function throws a warning of invalid year if the file with year in the name is not found. Also
 must be in the current working directory
}
\examples{
\dontrun{fars_summarize_years("2013")}
\dontrun{fars_summarize_years(2013)}
\dontrun{fars_summarize_years(c(2012, 2013, 2014, 2015))}
\dontrun{fars_summarize_years(c("2012", "2013", "2014", "2015"))}

}
