% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Names file with year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or numeric as year}
}
\value{
the function prints the static filename appended with year name passed as argument. The functions throws
an error when a non character or non numeric value is passed as an argument.
}
\examples{
make_filename(1976)
make_filename("1976")
make_filename(76)

}
